(editor) Convert allergens section to TypeScript
